Fix error when running on Linux
In Linux the capitalization of letters of files is important. It could
not find `hunterBotSpriteSheet`, because only `HunterBotSpriteSheet`
existed.

The current convention in the images folder also seems to be lowercase
for the first letter, which is why I chose to rename the file instead of
changing it in the code.